Crate mio_aio

MIO bindings for POSIX AIO

Structs

AioCb

LioCb

Enums

AioFsyncMode
Mode for AioCb::fsync. Controls whether only data or both data and metadata are synced.

Errno

LioError
Error types that can be returned by LioCb::submit

LioOpcode
When used with lio_listio, determines whether a given aiocb should be used for a read operation, a write operation, or ignored. Has no effect for any other aio functions.